SW-5370 - Add new component for many to many associations. Add if condition in the grid controller to remove the records directly if no destroy api defined. Remove many to many configuration and components from the detail window. Add pagination for the searchAssociationAction. Add new grid configuration to apply different column configuration without override any function.

+++ templates/_default/backend/base/application/Shopware.grid.Controller.js
@@ -353,6 +353,11 @@ Ext.define('Shopware.grid.Controller', {
      */
     onDeleteItem: function (grid, record) {
         var me = this;
+
+        if (!me.hasModelAction(record, 'destroy')) {
+            grid.getStore().remove(record);
+            return true;
+        }
         me.onDeleteItems(grid, [ record ], null);
     },
